requirements.txt
acme==2.6.0
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certbot==2.6.0
certbot-nginx==2.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
ConfigArgParse==1.5.3
configobj==5.0.8
cryptography==40.0.2
distlib==0.3.6
distro==1.8.0
Django==4.2.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-extra-fields==3.0.2
django-localflavor==4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
josepy==1.13.0
kombu==5.2.4
more-itertools==9.1.0
parsedatetime==2.6
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyRFC3339==1.1
python-decouple==3.8
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
pywin32==306
redis==4.5.5
requests==2.31.0
sentry-sdk==1.24.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
zipp==3.15.0
